requirements-basic.txt
beautifulsoup4==4.4.1
#decorator==4.0.10
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-crispy-forms==1.6.0
django-datetime-widget==0.9.3
django-djconfig==0.5.0
django-grappelli==2.8.1
#django-postgrespool==0.3.0
django-summernote==0.8.5
#ipdb==0.10.1
#ipython==4.2.1
#ipython-genutils==0.1.0
oauthlib==1.1.2
#pexpect==4.2.0
#pickleshare==0.7.2
Pillow==3.3.0
#pipdeptree==0.6.0
#psycopg2==2.6.1
#ptyprocess==0.5.1
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
#simplegeneric==0.8.1
#SQLAlchemy==1.0.14
static3==0.7.0
#traitlets==4.2.2
workdays==1.4
